// src/common/models/deep-Link.class.ts
export class DeepLinkObject {
    customerId: string;
    messageTitle: string
    messageBody: string
    messageGroupId: string;
    policyNumber: string;
    call2action: string;
    premiumDueDate: Date;
    paymentDate: string;
    claimId: string;
    statusDate: Date;
    status: string;
    messageType: string;
    fitsenseScreenPath: {}
    expirationDate: string
    topic: string
}